backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
playwright==1.39.0
beautifulsoup4==4.12.2
lxml==4.9.3
tldextract==3.4.4
html5lib==1.1
aiohttp==3.8.5
jinja2==3.1.2
weasyprint==60.1
loguru==0.7.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
sqlalchemy==2.0.22
alembic==1.12.0
